+++ 2D_Game_Engine/src/Timer/Timer.h
@@ -11,6 +11,6 @@ struct Timer
 		const Uint64 delta{ current_ticks - previous_ticks };
 		previous_ticks = current_ticks;
 		static const Uint64 TICKS_PER_SECOND{ SDL_GetPerformanceFrequency() };
-		elapsed_seconds = delta / static_cast<float>(TICKS_PER_SECOND);
+		elapsed_seconds = delta / static_cast<double>(TICKS_PER_SECOND);
 	}
 };
